+++ NEWS.md
@@ -8,6 +8,8 @@
 
 ## Other Changes
 
+* `predict()` can now be called on the workflow returned from `last_fit()` (#294, #295, #296).
+
 * tune now supports setting the `event_level` option from yardstick through the control objects (i.e. `control_grid(event_level = "second")`) (#240, #249).
 
 * tune now supports workflows created with the new `workflows::add_variables()` preprocessor.